Up to — Section 12, Lecture 122: "Testing User Interaction" <16:54> [00:51]
In general, don't like these tests much. They are checking for things that might work in tests, but not for the user. User using browser doesn't care if tests work, but doesn't work for them in browser. The last invalid amount is a good example of this being silly for end user as it's checking something that's not even possible for the user to do. It's just testing internal functions. Which is fine, just not as the only test. There should be a user behavior test using browser to be sure entering invalid amount isn't even possible as that is the case. So '12.122' isn't even possible to enter. It would end up just being '12.22'.